Addressed PR nits.
